requirements.txt
pandas
numpy
matplotlib
requests
python-dateutil
tqdm
python-dotenv
